OpenWith: MacOSAppProvider refactor20!
